+++ src/main/java/app/bottlenote/global/security/SecurityUtil.java
@@ -1,8 +1,6 @@
 package app.bottlenote.global.security;
 
 import app.bottlenote.global.security.customPrincipal.CustomUserContext;
-import app.bottlenote.user.repository.OauthRepository;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
@@ -10,11 +8,8 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 
 @Component
-@RequiredArgsConstructor
 @Slf4j
 public class SecurityUtil {
-
-	private final OauthRepository oauthRepository;
 
 	public static Long getCurrentUserId() {
 		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
@@ -27,5 +22,4 @@ public class SecurityUtil {
 
 		return userId;
 	}
-
 }
