+++ src/main/java/app/bottlenote/global/security/jwt/JwtAuthenticationFilter.java
@@ -5,14 +5,13 @@ import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.lang.NonNull;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -21,7 +20,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 	public static final String AUTHORIZATION_HEADER = "Authorization";
 	public static final String BEARER_PREFIX = "Bearer ";
 
-	private final JwtTokenProvider jwtTokenProvider;
+	private final JwtAuthenticationManager jwtAuthenticationManager;
 
 	@Override
 	protected void doFilterInternal(
@@ -33,7 +32,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 		String token = resolveToken(request);
 
 		if (isValidToken(token)) {
-			Authentication authentication = jwtTokenProvider.getAuthentication(token);
+			Authentication authentication = jwtAuthenticationManager.getAuthentication(token);
 			SecurityContextHolder.getContext().setAuthentication(authentication);
 		}
 
@@ -54,7 +53,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 	 * 추출된 토큰이 실제 유효한 토큰인지 검증.
 	 */
 	private boolean isValidToken(String token) {
-		return jwtTokenProvider.validateToken(token);
+		return JwtTokenValidator.validateToken(token);
 	}
 
 	/**
